chore: finish cleaning up old logger code
