+++ internal/logger/logger.go
@@ -3,7 +3,9 @@ package logger
 import (
 	"os"
 	"strings"
+	"time"
 
+	"github.com/gofiber/fiber/v2"
 	"github.com/sirupsen/logrus"
 )
 
@@ -23,6 +25,34 @@ func InitializeAndConfigure() {
 	configureLogLevel()
 
 }
+
+// Logger returns a middleware that logs HTTP requests
+func Logger() fiber.Handler {
+	return func(c *fiber.Ctx) error {
+		start := time.Now()
+
+		// Continue chain
+		err := c.Next()
+
+		// After request
+		stop := time.Now()
+		latency := stop.Sub(start)
+
+		// Log using logrus logger
+		InfoWithFields("Request", map[string]interface{}{
+			"timestamp": stop.Format("2006/01/02 - 15:04:05"),
+			"status":    c.Response().StatusCode(),
+			"latency":   latency,
+			"ip":        c.IP(),
+			"method":    c.Method(),
+			"path":      c.Path(),
+			"handler":   c.Route().Name,
+		})
+
+		return err
+	}
+}
+
 func configureLogLevel() {
 	log.SetLevel(logrus.InfoLevel)
